requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
twilio==9.3.7
requests==2.32.3
emergentintegrations
sqlalchemy==2.0.36